uint32 ngram counts in quant, some info statements on different format LM reading

+++ branches/sphinxbase-trie/src/libsphinxbase/lm/lm_trie_quant.h
@@ -52,12 +52,12 @@ uint8 lm_trie_quant_to_train(lm_trie_quant_t *quant);
 /**
  * Trains prob and backoff quantizer for specified ngram order on provided raw ngram list
  */
-void lm_trie_quant_train(lm_trie_quant_t *quant, int order, uint64 counts, lm_ngram_t* raw_ngrams);
+void lm_trie_quant_train(lm_trie_quant_t *quant, int order, uint32 counts, lm_ngram_t* raw_ngrams);
 
 /**
  * Trains only prob quantizer for specified ngram order on provided raw ngram list
  */
-void lm_trie_quant_train_prob(lm_trie_quant_t *quant, int order, uint64 counts, lm_ngram_t* raw_ngrams);
+void lm_trie_quant_train_prob(lm_trie_quant_t *quant, int order, uint32 counts, lm_ngram_t* raw_ngrams);
 
 /**
  * Writes specified weight for middle-order ngram. Quantize it if needed
